Update README and internals
- Documented `--auto` and `--data-dir` options in README.md
- Switched file handling from `int` to `size_t` for correct type usage
- Added default argument for `database_paths` in private rawtoaces_util_priv.h

+++ src/rawtoaces/main.cpp
@@ -44,11 +44,11 @@ int main( int argc, const char *argv[] )
     bool empty  = true;
     bool result = true;
 
-    int file_index  = 0;
-    int total_files = 0;
+    size_t file_index  = 0;
+    size_t total_files = 0;
 
     for ( auto const &batch: batches )
-        total_files += static_cast<int>( batch.size() );
+        total_files += batch.size();
 
     for ( auto const &batch: batches )
     {
